finalized routes by date. replaced add post form with a modal
