+++ public/js/myprofile.js
@@ -1,34 +1,3 @@
-// add a new post
-const addPost = async (event) => {
-    event.preventDefault();
-
-    const title = document.querySelector('#new-post-title').value.trim();
-    const content = document.querySelector('#new-post').value.trim();
-    const public = document.getElementById('new-post-public').checked;
-    const errorMessage = document.querySelector('.error');
-
-    if(content.length < 2 || title.length < 2) {
-        errorMessage.style.display = "block";
-        return;
-    };
-
-    const response = await fetch('/api/post', {
-        method: 'POST',
-        body: JSON.stringify({ title, content, public }),
-        headers: { 'Content-Type': 'application/json' }
-      });
-      
-      if (response.ok) {
-        document.location.reload();
-      } else {
-        alert('Post could not be added, please try again');
-      }
-    }
-
-document.querySelector('#newPostForm').addEventListener('submit', addPost);
-
-  
-
 
 // update a post
 const updatePost = async (event) => {
