+++ views/layouts/main.handlebars
@@ -1,10 +1,11 @@
+<!DOCTYPE html>
 <html lang="en">
 <head>
     <meta charset="UTF-8">
     <meta http-equiv="X-UA-Compatible" content="IE=edge">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <link rel="stylesheet" href="https://stackpath.bootstrapcdn.com/bootstrap/4.3.1/css/bootstrap.min.css">
-    <link rel="stylesheet" href="../css/style.css">
+    <link rel="stylesheet" href="/css/style.css">
     <title>The Happy Place</title>
 </head>
 <body>
@@ -15,9 +16,12 @@
             <a class="nav-link" href="/api/profile">My Profile</a>
             {{#if logged_in}}
             <a class="nav-link" id="logout" href="/">Logout</a>
+            <!-- Trigger/Open The Modal -->
+            <button id="modal-btn">New Post</button>
             {{else}}
             <a class="nav-link" href="/login">Login</a>
             {{/if}}
+
         </nav>
     </header>
 
@@ -34,11 +38,47 @@
         {{{ body }}}
     </main>
 
+    <!-- The Modal -->
+        <div id="instructions-modal" class="modal">
+            <!-- Modal content -->
+            <div class="modal-content">
+                <div class="modal-header">
+                    <span class="modal-close">&times;</span>
+                </div>
+                    {{!-- add a new post --}}
+                    <section class="modal-body">
+                        <h2 class="card-header">What are you Grateful for today?</h2>
+                        <form id="newPostForm">
+                            <div class="form-group">
+                                <label for="new-post-title">Title:</label>
+                                <input class="form-control" type="text" id="new-post-title" required aria-required="true"/>
+                            </div>
+                            <div class="form-group">
+                                <textarea name="new-post" id="new-post" class="comment-field" maxlength="250" required aria-required="true"></textarea>
+                            </div>
+                            <div class="form-group">
+                                <p>Public posts are visible to all viewers. Private posts are only visible to you.</p>
+                                    <input type="radio" id="new-post-public" name="visibility" value="true" checked>
+                                    <label for="public">Public</label><br>
+                                    <input type="radio" id="new-post-private" name="visibility" value="false">
+                                    <label for="private">Private</label><br>
+                            </div>
+                            <p class="error" style="display: none;">❌ Please enter a complete title and post content.</p>
+                            <button class="btn-primary" type="submit">Create Post</button>
+                        </form>
+                    </section>
+                <div class="modal-footer">
+                    <h3>Posting as <strong>{{username}}</strong></h3>
+                </div>
+            </div>
+        </div>
+
 
     {{#if logged_in}}
     <script src="/js/logout.js"></script>
     {{/if}}
-    <script src="./js/displayquote.js"></script>
-    <script src="./js/quotes.js"></script>
+    <script src="/js/newPostModal.js"></script>
+    <script src="/js/displayquote.js"></script>
+    <script src="/js/quotes.js"></script>
 </body>
 </html>
