+++ views/myprofile.handlebars
@@ -13,28 +13,6 @@
     <button class="btn-primary" data-user="{{user.name}}" id="bio-btn">Update Bio</button>
     <a href="mailto:{{user.email}}">{{user.email}}</a>
 </aside>
-{{!-- add a new post --}}
-<section>
-    <h2 class="card-header">New Post</h2>
-    <form id="newPostForm">
-        <div class="form-group">
-            <label for="new-post-title">Title:</label>
-            <input class="form-control" type="text" id="new-post-title" required aria-required="true"/>
-        </div>
-        <div class="form-group">
-            <textarea name="new-post" id="new-post" class="comment-field" maxlength="250" required aria-required="true"></textarea>
-        </div>
-        <div class="form-group">
-              <p>Public posts are visible to all viewers. Private posts are only visible to you.</p>
-                <input type="radio" id="new-post-public" name="visibility" value="true" checked>
-                <label for="public">Public</label><br>
-                <input type="radio" id="new-post-private" name="visibility" value="false">
-                <label for="private">Private</label><br>
-        </div>
-        <p class="error" style="display: none;">❌ Please enter a complete title and post content.</p>
-        <button class="btn-primary" type="submit">Create Post</button>
-    </form>
-</section>
 {{!-- My posts --}}
 <section>
     <h2>
